pyproject.toml
[tool.poetry]
name = "linus-kannadb"
version = "0.1.0"
description = "FEH database"
authors = ["Leenis, Poron, Avestis"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
alabaster = "^0.7.13"
appdirs = "^1.4.4"
argon2-cffi = "^23.1.0"
astroid = "^2.15.8"
atomicwrites = "^1.4.1"
attrs = "^23.1.0"
babel = "^2.12.1"
backcall = "^0.2.0"
beautifulsoup4 = "^4.12.2"
black = "^23.9.1"
certifi = "^2023.7.22"
cffi = "^1.15.1"
chardet = "^5.2.0"
click = "^8.1.7"
coreapi = "^2.3.3"
coreschema = "^0.0.4"
coverage = "^7.3.1"
decorator = "^5.1.1"
defusedxml = "^0.7.1"
django = "^4.2.5"
django-admin-view-permission = "^1.9"
django-advanced-filters = "^2.0.0"
django-allauth = "^0.57.0"
django-anymail = "^10.1"
django-braces = "^1.15.0"
django-cleanup = "^8.0.0"
django-clear-cache = "^0.3"
django-cloudinary-storage = "^0.3.0"
django-coverage-plugin = "^3.1.0"
django-crispy-forms = "^2.0"
django-debug-toolbar = "^4.2.0"
django-environ = "^0.11.2"
django-extensions = "^3.2.3"
django-model-utils = "^4.3.1"
django-redis = "^5.3.0"
django-storages = "^1.14"
djangorestframework = "^3.14.0"
docutils = "^0.20.1"
factory-boy = "^3.3.0"
faker = "^19.6.2"
flake8 = "^6.1.0"
gunicorn = "^21.2.0"
html5lib = "^1.1"
idna = "^3.4"
imagesize = "^1.4.1"
importlib-metadata = "^6.8.0"
ipdb = "^0.13.13"
ipython = "^8.15.0"
ipython-genutils = "^0.2.0"
isort = "^5.12.0"
itypes = "^1.2.0"
jedi = "^0.19.0"
jinja2 = "^3.1.2"
lazy-object-proxy = "^1.9.0"
lxml = "^4.9.3"
markupsafe = "^2.1.3"
mccabe = "^0.7.0"
more-itertools = "^10.1.0"
oauthlib = "^3.2.2"
packaging = "^23.1"
parso = "^0.8.3"
pathlib2 = "^2.3.7.post1"
pathspec = "^0.11.2"
pexpect = "^4.8.0"
pickleshare = "^0.7.5"
pillow = "^10.0.1"
pluggy = "^1.3.0"
prompt-toolkit = "^3.0.39"
psycopg2-binary = "^2.9.7"
ptyprocess = "^0.7.0"
py = "^1.11.0"
pycodestyle = "^2.11.0"
pycparser = "^2.21"
pycurl = "^7.45.2"
pyflakes = "^3.1.0"
pygments = "^2.16.1"
pylibmc = "^1.6.3"
pylint = "^2.17.6"
pyparsing = "^3.1.1"
pytest = "^7.4.2"
pytest-django = "^4.5.2"
pytest-sugar = "^0.9.7"
python-dateutil = "^2.8.2"
python-slugify = "^8.0.1"
python3-openid = "^3.2.0"
pytz = "^2023.3.post1"
redis = "^5.0.1"
regex = "^2023.8.8"
requests = "^2.31.0"
requests-oauthlib = "^1.3.1"
sentry-sdk = "^1.31.0"
simplejson = "^3.19.1"
six = "^1.16.0"
snowballstemmer = "^2.2.0"
soupsieve = "^2.5"
sphinx = "^7.2.6"
sphinxcontrib-websupport = "^1.2.6"
sqlparse = "^0.4.4"
termcolor = "^2.3.0"
text-unidecode = "^1.3"
toml = "^0.10.2"
tomli = "^2.0.1"
traitlets = "^5.10.1"
typed-ast = "^1.5.5"
typing-extensions = "^4.8.0"
unidecode = "^1.3.7"
uritemplate = "^4.1.1"
urllib3 = "^2.0.5"
w3lib = "^2.1.2"
wcwidth = "^0.2.6"
webencodings = "^0.5.1"
werkzeug = "^2.3.7"
whitenoise = "^6.5.0"
wrapt = "^1.15.0"
zipp = "^3.17.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py38']
include = '\.py$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''


[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 120
profile = "black"

[tool.ruff]
#Avoid enforcing line-length violations (`E501`)
ignore = ["E501"]